Made some changes, details below
Reformated a lot of the code to follow the 80 colum rule. Not
everywhere, though, as some of the sections looked better without it.

Made consistent asterisk placement for pointers. Currently it is on the
left side next to the type, but this may change in the future.

Took out <linux/limits.h> because it wasn't necessary and for
portability. Replaced it with a string concatination function.

Added function prototypes at the start of the file.

Made various changes.