Made some style changes, see below
Fixed the indentation on some lines where the four extra indentation
spaces for a continuation line were before the two tabs, not afterward.

Fixed how some of the lines are indented, some were over 79 characters
and some still are but were left in because they make sense.

Removed a lot of unnecessary vertical whitespace.

Removed code that had been commented out, if it needs to be recovered
then look in past commits.